confirm on delete

+++ src/pages/UserManagement.jsx
@@ -68,7 +68,12 @@ const UserManagement = () => {
             setDeleteResponse(response);
         }
 
+        if(window.confirm('Are you sure you want to delete?')){
+
         deleteUserFromDB();
+
+        }
+
         fetchUserDetails();
       }
     },
